Initial app with front end

+++ routes/clips.js
@@ -9,40 +9,76 @@ const { check, validationResult } = require("express-validator");
 // @access  Public
 router.get("/:id", async (req, res) => {
   try {
-    const clipData = await Clip.find({ id: req.params.id }).select("content");
-    res.json(clipData);
-    await Clip.findOneAndDelete({ id: req.params.id });
+    const clipData = await Clip.findOne({ id: req.params.id })
+      .select("content")
+      .select("expires");
+
+    if (clipData) {
+      if (
+        clipData.expires &&
+        new Date(clipData.expires).getTime() <= new Date().getTime()
+      ) {
+        await Clip.findOneAndDelete({ id: req.params.id });
+      } else if (!clipData.expires) {
+        await Clip.findOneAndDelete({ id: req.params.id });
+      }
+    }
+
+    res.send(clipData);
   } catch (err) {
-    console.error(err.message);
-    res.status(500).send("Server Error");
+    console.error(err);
+    res.status(500).json({ msg: "Server error occured!!" });
   }
 });
 
-// @route   POST api/clip/:id
-// @desc    Delete user content if already exist and Create new user content
+// @route   put api/clip/:id
+// @desc    Update user content if already exist else Create new user content
 // @access  Public
-router.post(
+router.put(
   "/:id",
-  [check("content", "Please include some content").isLength({ min: 1 })],
+  [
+    check("content", "Please include content").isLength({ min: 1 }),
+    check("expiresAfter", "Please include expires time in seconds").isLength({
+      min: 1,
+    }),
+  ],
   async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
       return res.status(400).json({ errors: errors.array() });
     }
 
-    const { content } = req.body;
+    const { content, expiresAfter, createdAt } = req.body;
     const id = req.params.id;
 
-    const clipData = new Clip({ id, content });
+    const clipData = {};
+    clipData.content = content;
+    clipData.createdAt = new Date();
+    clipData.expiresAfter = expiresAfter;
+
+    var created = new Date();
+    if (expiresAfter !== 0) {
+      clipData.expires = created.setSeconds(
+        created.getSeconds() + expiresAfter
+      );
+    }
 
     try {
-      await Clip.findOneAndDelete({ id: id });
-      await clipData.save();
+      await Clip.findOneAndUpdate(
+        { id: id },
+        { $set: clipData },
+        {
+          new: true,
+          upsert: true,
+        }
+      );
+      // await clipData.save();
 
-      return res.status(201).json({ msg: "Clip created" });
+      return res.json({ msg: "Clip created" });
     } catch (err) {
-      console.error(err.message);
-      res.status(500).send("Server Error");
+      console.log("Error: " + err);
+      console.error("Err msg:" + err.message);
+      res.status(500).json({ msg: "Server error occured!!" });
     }
   }
 );
